Updates and fixes

+++ main/update_db.py
@@ -12,46 +12,48 @@ def update_db(data_source):
     data_source = 2 - champion.gg
     '''
     if data_source == 1:
-        # Delete data from db
-        Champ_winrate.objects.filter(source=data_source).all().delete()
-
         url = 'https://www.metasrc.com/5v5/stats'
         webpage = requests.get(url)
         webpage_preety = soup(webpage.content, 'html.parser')
         champ_list = webpage_preety.find_all("tr", {"class": "_sbzxul"})
 
-        for champ in champ_list:
-            try:
-                data = champ.find_all("td")
-                champ_name = data[0].find_next("span").get_text()
-                champ_lane = data[1].find_next("div").get_text().lower()
-                champ_win = data[5].get_text()[:-1]
-                champ = Champ_winrate(name= champ_name, role = champ_lane, win_rate = champ_win, source = 1)
-                champ.save()
-            except:
-                return -1
+        if len(champ_list) > 100:
+            # Delete data from db
+            Champ_winrate.objects.filter(source=data_source).all().delete()
+
+            for champ in champ_list:
+                try:
+                    data = champ.find_all("td")
+                    champ_name = data[0].find_next("span").get_text()
+                    champ_lane = data[1].find_next("div").get_text().lower()
+                    champ_win = data[5].get_text()[:-1]
+                    champ = Champ_winrate(name= champ_name, role = champ_lane, win_rate = champ_win, source = 1)
+                    champ.save()
+                except:
+                    pass
 
     if data_source == 2:
-        # Delete data from db
-        Champ_winrate.objects.filter(source=data_source).all().delete()
-
         url = 'https://champion.gg/tierlist'
         webpage = requests.get(url)
         webpage_preety = soup(webpage.content, 'html.parser')
 
         champ_list = webpage_preety.find_all("div", {"class": "fYmnzJ"})
 
-        for champ in champ_list:
-            try:
-                data = champ.find_all("div")
-                champ_name = data[3].get_text()
-                champ_lane = data[0].get_text()[5:]
-                champ_win = data[2].find_next("span").get_text()[:-1]
-                
-                champ = Champ_winrate(name= champ_name, role = champ_lane, win_rate = champ_win, source = 2)
-                champ.save()
-            except:
-                return -1
+        if len(champ_list) > 100:
+            # Delete data from db
+            Champ_winrate.objects.filter(source=data_source).all().delete()
+
+            for champ in champ_list:
+                try:
+                    data = champ.find_all("div")
+                    champ_name = data[3].get_text()
+                    champ_lane = data[0].get_text()[5:]
+                    champ_win = data[2].find_next("span").get_text()[:-1]
+                    
+                    champ = Champ_winrate(name= champ_name, role = champ_lane, win_rate = champ_win, source = 2)
+                    champ.save()
+                except:
+                    pass
 
     '''
     Select champion names from database and save to file
@@ -64,5 +66,3 @@ def update_db(data_source):
     with open(os.path.join(dirname, 'champ_list.txt'), 'w') as file:
         for champ in champs:
             file.write(champ+'\n')
-
-    return 0
